Organization and Sprite Flipping
Confirmed controls work with gamepad, organized assets folder, and made
character face the direction its moving.
