+++ Super Broomball 2016/Assets/Sprites.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 0ea0ffe8ab7c8b24c9ff86f064165c76
+folderAsset: yes
+timeCreated: 1455895826
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
