+++ Super Broomball 2016/Super Broomball 2016.CSharp.csproj
@@ -63,8 +63,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Assets\PlayerScript.cs" />
-    <Compile Include="Assets\SpinThisThing.cs" />
+    <Compile Include="Assets\_Scripts\PlayerScript.cs" />
+    <Compile Include="Assets\_Scripts\SpinThisThing.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\SyntaxTree\UnityVS\2015\UnityVS.CSharp.targets" />
 </Project>
